refactor: move web utils to `ui/react/internal/utils` (#1716)
