+++ src/core/utils/convertSnakeToCamelCase.test.ts
