+++ src/core/utils/convertSnakeToCamelCase.ts
