+++ src/fund/utils/getFundingPopupSize.ts
@@ -1,7 +1,7 @@
 import {
   getWindowDimensions,
   type popupSizes,
-} from '../../internal/utils/getWindowDimensions';
+} from '../../ui/react/internal/utils/getWindowDimensions';
 import { ONRAMP_POPUP_HEIGHT, ONRAMP_POPUP_WIDTH } from '../constants';
 import { ONRAMP_BUY_URL } from '../constants';
