+++ src/wallet/components/WalletDropdownFundLink.tsx
@@ -2,8 +2,8 @@ import { useCallback, useMemo } from 'react';
 import { useIcon } from '../../core-react/internal/hooks/useIcon';
 import { useGetFundingUrl } from '../../fund/hooks/useGetFundingUrl';
 import { getFundingPopupSize } from '../../fund/utils/getFundingPopupSize';
-import { openPopup } from '../../internal/utils/openPopup';
 import { cn, color, pressable, text as themeText } from '../../styles/theme';
+import { openPopup } from '../../ui/react/internal/utils/openPopup';
 import type { WalletDropdownFundLinkReact } from '../types';
 
 export function WalletDropdownFundLink({
